+++ src/Pet4U.API/Controllers/ApplicationController.cs
@@ -1,9 +1,4 @@
 using Microsoft.AspNetCore.Mvc;
-using Pet4U.API.Extensions;
-using Pet4U.Application.UseCases.CreateVolunteer;
-using Pet4U.Domain;
-using Pet4U.Domain.Modules;
-using Pet4U.Domain.Shared;
 using Pet4U.Response;
 
 namespace Pet4U.API;
