+++ src/Pet4U.API/Response/Envelope.cs
@@ -1,7 +1,4 @@
-using Pet4U.Domain.Shared;
-
 namespace Pet4U.Response;
-
 
 //public record ResponseError(Error error, string?  InvalidField);
