+++ src/Pet4U.Application/UseCases/CreateVolunteer/Handler.cs
@@ -1,6 +1,6 @@
-using Pet4U.Domain.Modules;
-using Pet4U.Domain;
-using Pet4U.Domain.Shared;
+using Pet4U.Domain.Ids;
+using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Volunteers;
 
 namespace Pet4U.Application.UseCases.CreateVolunteer;
