+++ src/Pet4U.Application/UseCases/CreateVolunteer/ICreateVolunteerHandler.cs
@@ -1,4 +1,5 @@
-using Pet4U.Domain.Shared;
+
+using Pet4U.Domain.ValueObjects;
 
 namespace Pet4U.Application.UseCases.CreateVolunteer;
