+++ src/Pet4U.Domain/Species/Species.cs
@@ -1,6 +1,7 @@
-using Pet4U.Domain.Shared;
+using Pet4U.Domain.Ids;
+using Pet4U.Domain.ValueObjects;
 
-namespace Pet4U.Domain.Modules;
+namespace Pet4U.Domain.Species;
 
 public class Species : Entity<SpeciesId>
 {
