+++ src/Pet4U.Domain/ValueObjects/Description.cs
@@ -1,4 +1,6 @@
-namespace Pet4U.Domain.Shared;
+using Pet4U.Domain.ValueObjects;
+
+namespace Pet4U.Domain.ValueObjects;
 
 public record Description
 {
