+++ src/Pet4U.Domain/ValueObjects/Entity.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain.Shared;
+namespace Pet4U.Domain.ValueObjects;
 
 public abstract class Entity<T> where T : notnull
 {
