+++ src/Pet4U.Domain/ValueObjects/Errors.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain.Shared;
+namespace Pet4U.Domain.ValueObjects;
 
 public static class Errors
 {
