+++ src/Pet4U.Domain/ValueObjects/FullName.cs
@@ -1,7 +1,8 @@
-using Pet4U.Domain.Shared;
-using static Pet4U.Domain.Shared.Errors;
 
-namespace Pet4U.Domain
+using Pet4U.Domain.ValueObjects;
+using static Pet4U.Domain.ValueObjects.Errors;
+
+namespace Pet4U.Domain.ValueObjects
 {
   public record FullName
   {
@@ -20,7 +21,7 @@ namespace Pet4U.Domain
     public static Result<FullName> Create(string firstName, string lastName, string middleName)
     {
       if(string.IsNullOrWhiteSpace(firstName))
-        return Errors.General.ValueIsInvalid("first name");//Error.Validation("value.is.invalid", "Invalid first name wasn't accepted");
+        return General.ValueIsInvalid("first name");//Error.Validation("value.is.invalid", "Invalid first name wasn't accepted");
 
       if(firstName.Length > Constants.MAX_LOW_TEXT_LENGTH)
         return General.LengthIsInvalid(firstName);
