+++ src/Pet4U.Domain/ValueObjects/Result.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain.Shared;
+namespace Pet4U.Domain.ValueObjects;
 
 public class Result
 {
