+++ src/Pet4U.Domain/Volunteers/SocialNetwork.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain
+namespace Pet4U.Domain.Volunteers
 {
   public record SocialNetwork
   {
