+++ src/Pet4U.Domain/Volunteers/Status.cs
@@ -1,4 +1,4 @@
-namespace Pet4U.Domain
+namespace Pet4U.Domain.Volunteers
 {
   public enum Status {NeedHelp, LookingForHome, FoundHome};
 }
