+++ src/Pet4U.Infrastructure/Configurations/VolunteerConfiguration.cs
@@ -3,12 +3,13 @@ using System.Reflection.Metadata;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 using Pet4U.Domain;
-using Pet4U.Domain.Modules;
-using Pet4U.Domain.Shared;
+using Pet4U.Domain.Ids;
+using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Volunteers;
 
 namespace Pet4U.Infrastructure 
 {
-    public class VolunteerConfiguration : IEntityTypeConfiguration<Pet4U.Domain.Modules.Volunteer>
+    public class VolunteerConfiguration : IEntityTypeConfiguration<Volunteer>
     {
         public void Configure(EntityTypeBuilder<Volunteer> builder)
         {
