+++ src/Pet4U.Infrastructure/Repositories/VolunteersRepository.cs
@@ -1,7 +1,8 @@
 using Microsoft.EntityFrameworkCore;
 using Pet4U.Application.UseCases.CreateVolunteer;
-using Pet4U.Domain.Modules;
-using Pet4U.Domain.Shared;
+using Pet4U.Domain.Ids;
+using Pet4U.Domain.ValueObjects;
+using Pet4U.Domain.Volunteers;
 
 namespace Pet4U.Infrastructure.Repositories;
